test was used to determine the speed of inverse algorthm. Present speed (for a 20k *1 inverse) is 1 block /12s. That's for a 20k*20k matrix it needs  about 67h.

+++ mamu/simplemm.cl
@@ -81,7 +81,7 @@ __kernel void mamu(__global float* a,
      * Compute element c[rowIndex, colIndex].
      **/
 
-    int sum = 0;
+    float sum = 0;
     for(int k = 0; k < K; k++){
         sum += a[rowIndex*K + k] * b[k*N + colIndex];
     }
